+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/serialization/FileExtension.kt
@@ -23,6 +23,7 @@ enum class FileExtension(
     JAVA(".java"),
     JAVASCRIPT(".js", setOf(".cjs", ".mjs")),
     KOTLIN(".kt"),
+    OBJECTIVE_C(".m"),
     PYTHON(".py"),
     RUBY(".rb"),
     SWIFT(".swift"),